Add files via upload
This project is a feature-rich mobile application developed using the Flutter-Dart framework and Firebase backend. It serves as a powerful SignUp/Login system for Android apps, allowing for real-time user data management and secure user authentication. This solution is highly customizable and can be seamlessly integrated into a wide range of Android applications, providing users with a reliable and user-friendly experience.